requirements.txt
aiogram
aiosqlite
asyncopoisk
SQLAlchemy
python-decouple